pyproject.toml
[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = ["F841"]


[tool.ruff.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Path",
    "fastapi.Form",
    "fastapi.File"
]


[tool.mypy]
strict = true

[tool.poetry]
name = "code-quality-inspector"
version = "0.1.0"
description = ""
authors = ["Zoltan Dominguez <zoltan.dominguez@gmail.com>"]
readme = "README.md"
packages = [{include = "cqi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "~0.97.0"
httpx = "*"
python-multipart = "*"
uvicorn = "~0.22.0"
xmltodict = "~0.13.0"
boto3 = ">=1.26.129"
botocore = "*"
PyGithub = "~1.58.1"
tomli = "~2.0.1"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
pylint = "*"
ruff = "*"
semgrep = "*"
mypy = "*"
types-boto3 = "*"
boto3-stubs-lite = { version = "*", extras = ["dynamodb"] }
types-xmltodict = "*"

pytest = "*"
pytest-cov = "*"
pytest-html = "*"
pytest-random-order = "*"
requests = "*"
coverage = "*"
